+++ acme/client_config.go
@@ -10,7 +10,7 @@ import (
 	"os"
 	"time"
 
-	"github.com/xenolf/lego/certificate/certcrypto"
+	"github.com/xenolf/lego/certcrypto"
 	"github.com/xenolf/lego/registration"
 )
