+++ cmd/cmd_renew.go
@@ -8,8 +8,8 @@ import (
 	"time"
 
 	"github.com/urfave/cli"
+	"github.com/xenolf/lego/certcrypto"
 	"github.com/xenolf/lego/certificate"
-	"github.com/xenolf/lego/certificate/certcrypto"
 	"github.com/xenolf/lego/log"
 )
